将Resolver做了一层抽象
