+++ WF-Public/src/main/java/org/radrso/workflow/exec/StepExecutor.java
@@ -3,7 +3,7 @@ package org.radrso.workflow.exec;
 import org.radrso.workflow.entities.wf.WorkflowInstance;
 import org.radrso.workflow.exec.base.BaseStepAction;
 import org.radrso.workflow.exec.base.StepSubscriber;
-import org.radrso.workflow.resolvers.WorkflowResolver;
+import org.radrso.workflow.resolvers.BaseWorkflowConfigResolver;
 import rx.Observable;
 import rx.schedulers.Schedulers;
 
@@ -11,7 +11,7 @@ import rx.schedulers.Schedulers;
  * Created by raomengnan on 17-1-17.
  */
 public class StepExecutor {
-    public static void execute(BaseStepAction stepAction, WorkflowResolver workflowResolver){
+    public static void execute(BaseStepAction stepAction, BaseWorkflowConfigResolver workflowResolver){
         workflowResolver.getWorkflowInstance().setStatus(WorkflowInstance.RUNNING);
         Observable.just(workflowResolver)
                 .subscribeOn(Schedulers.io())
