+++ WF-Public/src/main/java/org/radrso/workflow/persistence/BaseWorkflowSynchronize.java
@@ -5,7 +5,6 @@ import org.radrso.workflow.entities.config.items.Step;
 import org.radrso.workflow.entities.response.WFResponse;
 import org.radrso.workflow.entities.wf.WorkflowErrorLog;
 import org.radrso.workflow.entities.wf.WorkflowInstance;
-import org.radrso.workflow.resolvers.WorkflowResolver;
 
 import java.util.List;
