+++ WF-Public/src/test/java/org/radrso/workflow/resolvers/TestStepResolver.java
@@ -5,14 +5,41 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.radrso.workflow.entities.config.items.Step;
 import org.radrso.workflow.entities.response.WFResponse;
-import org.radrso.workflow.resolvers.StepExecuteResolver;
+import org.radrso.workflow.resolvers.impl.StepExecuteResolver;
 
 /**
  * Created by rao-mengnan on 2017/3/12.
  */
 public class TestStepResolver {
+
+    public static String sum_1(String str, Integer... args){
+        int sum = 0;
+        for (int i = 0; i < args.length; i++)
+            sum += args[i];
+        return str + sum;
+    }
+
+    public static String sum_2(String str, int... args){
+        int sum = 0;
+        for (int i = 0; i < args.length; i++)
+            sum += args[i];
+        return str + sum;
+    }
+
     @Test
     public void testClassRequest() {
+        Step step = new Step();
+        step.setCall("class:org.radrso.workflow.resolvers.TestStepResolver");
+        step.setMethod("sum_1");
+        String[] paramNames = new String[]{"default", "default", "default", "default", "default"};
+        Integer[] params = new Integer[]{1, 2, 3, 4, 5};
+        StepExecuteResolver resolver = new StepExecuteResolver(step, new Object[]{"test:", params}, paramNames);
+        WFResponse response = resolver.classRequest();
+        System.out.println("Result:" + response);
+
+        step.setMethod("sum_2");
+        response = resolver.classRequest();
+        Assert.assertNotEquals(response.getCode(), 200);
     }
 
     @Test
@@ -40,5 +67,7 @@ public class TestStepResolver {
         response = resolver.netRequest();
         System.out.println(response);
         Assert.assertEquals(response.getMsg().contains("a14afef0f66fcffce3e0fcd2e34f6ff4"), true);
+
+
     }
 }
