+++ WF-Service/src/main/java/org/radrso/workflow/wfservice/executor/InstanceJobRunner.java
@@ -1,6 +1,6 @@
 package org.radrso.workflow.wfservice.executor;
 
-import org.radrso.workflow.resolvers.WorkflowResolver;
+import org.radrso.workflow.resolvers.BaseWorkflowConfigResolver;
 import org.radrso.workflow.entities.response.WFResponse;
 
 /**
@@ -8,6 +8,6 @@ import org.radrso.workflow.entities.response.WFResponse;
  */
 public interface InstanceJobRunner {
 
-    WFResponse startExecute(WorkflowResolver workflowResolver);
+    WFResponse startExecute(BaseWorkflowConfigResolver workflowResolver);
 
 }
